Moves MockFileSystem constructor performance test to benchmarks project
