+++ benchmarks/System.IO.Abstractions.Benchmarks/Program.cs
@@ -1,12 +1,13 @@
 ﻿namespace System.IO.Abstractions.Benchmarks
 {
     using BenchmarkDotNet.Running;
+    using System.Reflection;
 
     class Program
     {
         public static void Main(string[] args)
         {
-            BenchmarkRunner.Run<FileSystemAbstractionBenchmarks>();
+            BenchmarkRunner.Run(typeof(Program).Assembly);
         }
     }
 }
