+++ benchmarks/System.IO.Abstractions.Benchmarks/System.IO.Abstractions.Benchmarks.csproj
@@ -14,6 +14,7 @@
 	</PropertyGroup>
 	<ItemGroup>
 		<ProjectReference Include="../../src/System.IO.Abstractions/System.IO.Abstractions.csproj" />
+		<ProjectReference Include="../../src/System.IO.Abstractions.TestingHelpers/System.IO.Abstractions.TestingHelpers.csproj" />
 	</ItemGroup>
 	<ItemGroup>
 		<PackageReference Include="Microsoft.NETFramework.ReferenceAssemblies.net461" Version="1.0.0">
